# Faster engine that executes test cases all at once