--- a/engine/run_py.py
+++ b/engine/run_py.py
@@ -5,10 +5,10 @@ import importlib
 import tracemalloc
 
 run_id = os.path.basename(__file__).split('.')[0]
-
 input_pickle = '{:s}.input.pickle'.format(run_id)
 
 user_module = importlib.import_module(run_id)
+
 with open(input_pickle, mode='rb') as f:
     input_tuples = pickle.load(f)
 
@@ -18,7 +18,7 @@ for i, input_tuple in enumerate(input_tuples):
     tracemalloc.start()
     t1 = time.time()
 
-    # $FUNCTION_NAME will be replaced by the name of the user's function before this script is run.
+    # $FUNCTION_NAME will be replaced by the name of the user's function by the PythonRunner before this script is run.
     user_output = user_module.$FUNCTION_NAME(*input_tuple)
 
     t2 = time.time()