This PR changes the engine to run all test cases at once, so it's faster! Not much faster, maybe twice as fast so most problems are checked in <1 second.

```
# ada @ lovelace-dev-vm in ~/lovelace/lovelace-engine on git:faster-engine x [9:06:39] 
$ make test
Make sure the engine is running!
Running tests...
cd engine ; /usr/local/bin/python3 -m unittest test_api.py
Submitting ../solutions/python/chaos.py... 0.509199 seconds.
Submitting ../solutions/python/almost_pi.py... 1.051957 seconds.
Submitting ../solutions/python/babylonian_square_roots.py... 0.575816 seconds.
Submitting ../solutions/python/blood_types.py... 0.624332 seconds.
Submitting ../solutions/python/caesar_cipher.py... 0.452541 seconds.
Submitting ../solutions/python/colorful_resistors.py... 0.581917 seconds.
Submitting ../solutions/python/correlation_does_not_imply_causation.py... 0.710109 seconds.
Submitting ../solutions/python/definite_integrals.py... 0.778478 seconds.
Submitting ../solutions/python/dna_transcription.py... 0.590514 seconds.
Submitting ../solutions/python/earthquake_epicenters.py... 0.507587 seconds.
Submitting ../solutions/python/habitable_exoplanets.py... 0.459031 seconds.
Submitting ../solutions/python/nand_gate.py... 0.471494 seconds.
Submitting ../solutions/python/rocket_science.py... 0.539883 seconds.
Submitting ../solutions/python/wind_chill.py... 0.577333 seconds.
Submitting ../solutions/python/el_nino_intensities.py... 0.594903 seconds.
Submitting ../solutions/python/molecular_mass_calculator.py... 0.819335 seconds.
Submitting ../solutions/python/game_of_life.py... 2.712893 seconds.
Submitting ../solutions/python/ada_lovelaces_note_g.py... 1.850410 seconds.
Submitting ../solutions/python/flight_paths.py... 1.105991 seconds.
.
----------------------------------------------------------------------
Ran 1 test in 15.525s

OK
```

Resolves #25 
Resolves #36