@ali-ramadhan 2019-07-18
The runners take 1 input pickle but produce N output pickles where N is the number of test cases. This is slower as multiple files need to be pulled but still pretty fast as pulling files is fast compared to running python multiple times. The reason for this is that it could be a good way of querying the Linux container for how many test cases have been completed (could be done via `ls`). We can use this info to show the user a progress bar via AJAX, or even update test cases as they run.