@ali-ramadhan 2019-07-18
To make the JavaScript runner faster we'll have to modify the `glue_code` to execute all test cases at once, but my JavaScript is rusty and current performance isn't bad.