@ali-ramadhan 2019-07-18
Actually never mind, glue code is appended so we just keep appending then run `node` once.

Javascript performance is pretty good!

```
# ada @ lovelace-dev-vm in ~/lovelace/lovelace-engine on git:faster-engine x [10:25:17] 
$ make test                                                         
Make sure the engine is running!
Running tests...
cd engine ; /usr/local/bin/python3 -m unittest test_api.py
Submitting ../solutions/javascript/chaos.js... 0.725899 seconds.
Submitting ../solutions/javascript/colorful-resistors.js... 1.031705 seconds.
Submitting ../solutions/javascript/dna-transcription.js... 0.561259 seconds.
Submitting ../solutions/javascript/earthquake-epicenters.js... 0.687736 seconds.
Submitting ../solutions/javascript/rocket-science.js... 0.719510 seconds.
.
----------------------------------------------------------------------
Ran 1 test in 3.729s

OK
```