@ali-ramadhan 2019-07-18
Julia is much faster now but still ~10x slower than Python.

To really speed it up, we probably need to use glue code like in `run_js.py` and just call `julia` from the command line instead of initializing it through Python. Will be nice as we won't have to rely on the `julia` Python package, but we'll have to communicate between Python and Julia via files of some kind...

```
# ada @ lovelace-dev-vm in ~/lovelace/lovelace-engine on git:faster-engine x [10:46:17] 
$ make test
Make sure the engine is running!
Running tests...
cd engine ; /usr/local/bin/python3 -m unittest test_api.py
Submitting ../solutions/julia/rocket_science.jl... 9.461537 seconds.
Submitting ../solutions/julia/habitable_exoplanets.jl... 9.274049 seconds.
Submitting ../solutions/julia/nand_gate.jl... 9.117524 seconds.
Submitting ../solutions/julia/wind_chill.jl... 9.285537 seconds.
.
----------------------------------------------------------------------
Ran 1 test in 37.141s

OK
```